Fix: Use integer HP values for encounter completion logic
- Changed player_total_hp and monster_total_hp from f64 to u32 integers
- Updated HP comparison from <= 0.0 to == 0 for integer values
- Removed epsilon logic since we're now using exact integer comparisons
- This should fix the premature combat termination issue where K=33>0 but combat was stopping early

+++ simulation-wasm/src/context.rs
@@ -36,8 +36,8 @@ pub struct TurnContext {
 pub struct CombattantState {
     pub id: String,
     pub base_combatant: Combattant,
-    pub current_hp: f64,
-    pub temp_hp: f64,
+    pub current_hp: u32,
+    pub temp_hp: u32,
     pub conditions: Vec<CreatureCondition>,
     pub concentration: Option<String>, // ID of spell/concentration source
     pub position: Option<String>,      // Simplified position for future expansion
